Mount `/dev/shm` as shared mount between containers in pod
K8s mounts `/dev/shm` as a `bind` mount pointing to the same location for
all containers in a pod, so it can be shared amoung them. This causes
gVisor with VFS2 to mount `/dev/shm` over 9p to the gofer. This confuses
glibc which looks for a tmpfs mount for the directory to use with shm_open(3).

The shim now detects this intent and add makes the corresponding changes to
the spec to make `/dev/shm` be mounted as a pod shared tmpfs volume. It consists
of annotations to indicate that the volume is pod shared, and changes to mount
type from `bind` to `tmpfs`.

It also required changes in runsc to allow shared mounts with more restrictive
options, so it doesn't break containers that mount `/dev/shm` with less
restrictive options, like `ro`.

Closes #6943

PiperOrigin-RevId: 417508549